LayoutTests/http/tests/media/hls/hls-video-resize-expected.txt


Test <video> element initial resize event for HLS.

EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(resize)
Receive first resize event before loadedmetadata. OK
EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(loadedmetadata)
Receive loadmetadata event before second resize. OK
EVENT(resize)
EXPECTED (video.videoWidth == '640') OK
EXPECTED (video.videoHeight == '480') OK
END OF TEST
